feat: Decode static instructions
